update of files and from js format to jsx
