+++ ape_app/src/pages/components/cardImageContainer.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const CardImageContainer = ({children}) => {
     return(
-        <div className='cardImageContainer w-full h-[170px] relative'>
+        <div className='cardImageContainer flex w-full h-[170px] relative'>
             {children}
         </div>
     )
