+++ ape_app/src/pages/components/gliderControls.jsx
