+++ ape_app/src/pages/components/homePageBanner.jsx
