+++ ape_app/src/pages/components/mainSection.jsx
@@ -1,0 +1,21 @@
+import React from 'react'
+import TournamentListing from './tournamentListing';
+import TournamentScroller from './tournamentScroller';
+
+const MainSection = ({children}) => {
+  return (
+    <TournamentListing>
+        <section className=' bg-white'>
+            <div className=' pt-12 mx-4 my-0 overflow-hidden max-w-[1260px]'>
+                <TournamentScroller>
+                    <div role='group' aria-labelledby='18'>
+                        {children}
+                    </div>
+                </TournamentScroller>
+            </div>
+        </section>
+    </TournamentListing>
+  )
+}
+
+export default MainSection;
