+++ ape_app/src/pages/components/registrationClose.jsx
