+++ ape_app/src/pages/components/styleCard.jsx
