+++ ape_app/src/pages/eventos/create/index.jsx
@@ -1,0 +1,13 @@
+import NavBar from '../../navbar/navbar';
+import React from 'react';
+
+const createEvents = () => {
+    return(
+        <div className="box-border *:before:*:after">
+            <NavBar/>
+            
+        </div>
+    )
+}
+
+export default createEvents;
