+++ ape_app/src/pages/eventos/partidos/index.jsx
@@ -1,0 +1,11 @@
+import NavBar from '../../navbar/navbar';
+import React from 'react'
+
+const resultEvents = () => {
+    return(
+        <div className="box-border *:before:*:after">
+            <NavBar/>
+        </div>
+    )
+}
+export default resultEvents;
