+++ ape_app/src/pages/eventos/search/index.jsx
@@ -1,0 +1,11 @@
+import NavBar from '../../navbar/navbar';
+import React from 'react'
+
+const searchEvents = () => {
+    return(
+        <div className="box-border *:before:*:after">
+            <NavBar/>
+        </div>
+    )
+}
+export default searchEvents;
