+++ ape_app/src/pages/profile/edit/index.jsx
@@ -1,0 +1,11 @@
+/*Edit Profile page*/ 
+import NavBar from "../../navbar/navbar";
+import React from "react";
+
+const editProfile = () => {
+    return(
+        <div>editProfile</div>
+    )
+}
+
+export default editProfile;
