+++ ape_app/src/pages/profile/index.jsx
@@ -1,0 +1,11 @@
+/*View Profile page*/
+import React from "react";
+import NavBar from "../navbar/navbar";
+
+const viewProfile = () => {
+    return(
+        <div>viewProfile</div>
+    )
+}
+
+export default viewProfile;
